refactor: split service logic and api end point
